Change and break apart data structures
Instead of having a complete separate list of strings for seen bird names for
each user, we separate it out into a list of birds, a list of people, and a list
of sightings.

Each entry in the list of sightings references a bird and a person by ID, so if
a person has seen a specific bird, we just add an entry into the big sighting
list.

Included is a convert.py script and a remote_convert.sh to convert data stored
in the previous fashion to the new fashion, which is to be run before starting
the application again with the new model.