Add basic outline for page design
The basic page design is to have a bird search bar at the top with a login
button to the right. Some functionality of the site will be usable without
being logged in, but if logged in, you should be able to just quickly look up a
bird and then click "Sighted" to add it as a sighting.