Adjust duplication of menu items in the two menus
Previously, it was noticable that the sidebar menu was populated with the menu
items through JavaScript because it loaded a bit slow. With this commit, the
sidebar menu now have all the menu items originally, and then the collapsable
menu are just populated with copies of the items in the sidebar after the page
has loaded. This is a bit better since the collapsable menu won't ever be
expanded directly when loaded anyway, so you will hopefully never see that it
is initially empty.