Improve and fix bugs with registering new account
Now, you will stay and the register page both when creating and when failing to
create a new account, so that you can read the outcome of the creation. Also,
there was a bug which let the user create an account with invalid credentials,
which of course weren't permitted when trying to log in, so the user could never
log in.