Show sightings for person of logged in account
Previously, the flask session only stored the username, and then this username
was used to search for a person by name. But if the username and the account's
person's name were different, it failed. This has been fixed in this commit.
Now the user account's ID is stored in the session instead. And using this ID,
the account and the account's person ID are retrieved when necessary instead.