Fill birds in database and adjust sighting logic
With this commit, a lot of changes has been made. One of the biggest changes is
that the database is now populated with a complete list of birds, by their
binomial name, which can be searched in using the web page search bar. This
also means, that the functionality for the end user to add their own birds when
they added a new sighting, has been removed. Now, instead, they must search the
web page for a specific bird, and then just click the 'Sighted' button.