Rework account registration with email links
In order to avoid people from spamming new accounts, a user must now first pro-
vide an email address to get a link to the registration form where they will
input their desired username and password. This also means that, in the future,
it will be quite trivial to implement functionality to reset a user's password
by requesting a special link sent as an email.

These emails are dispatched using a real mail server if the MAIL_SERVER environ-
ment variable is set in the birding-web-service container. If it isn't, the
dispatched mails will instead just get dumped to the flask log.