Break apart app.py to use factory and blueprints
Now, the app is constructed by the function create_app in __init__.py file
instead. Also, in there, some functionality has been broken out to be put into
blueprints that are registered on the app instead. The structure is still not
ideal, but it is a step in the right direction.