Improve bird search result with thumbnails
With this commit, the logic for constructing the result items sent to the page
rendering has been moved to a new module called search_view. This new module has
been given a few very simple tests, and all of this results in a very clean
blueprint_search module which only performs searches and then invokes methods in
the view module to construct the items that are passed to rendering.