Fix bug with BirdSearchViewFactory
The factory used only the pictures and bird_thumbnail rows from the database
present when the factory was created. This caused some instances of the
application not to use the latest added pictures and thumbnails present in the
database.

With this commit, the factory now uses a builder, which is given a fresh set of
bird thumbnails and pictures rows each time the factory is invoked to create
search result items.