Fix issues with account registration
The account registration failed all the time because the readonly email input
didn't have an actual value, so the sanity check on server side couldn't
validate the token against the email correctly with what existed in the
database. Also, the account created flash shown when it works, has been
improved.