Break out logic from post_register_form with tests
There were a lot of logic when finalizing a user account registration that lied
in the authentication blueprint module that was better suited straight up in the
plain authentication module. With this commit, this logic has been moved there,
refactored, and supported with a few more unit tests.

It is still not optimal, but it is a step in the right direction.