Improve search result query matches
Now, search result matches that contain the sought after name query are
attributed correct query match ratios based on how much of their name actually
matched with the search. This will put 'Magpie' on top of the search result
because it perfectly matched, where previously it would be further down.