Add password reset functionality
Now, on the login page, there is a link to a password reset page where the user
can submit their email. A link to their password reset form is then sent to
them and when they follow this link thay are prompted to input a new password.
When this has been submitted, the password changes and they are brought back to
the login page.