Fix sightings home page crash when sightings exist
The sightings home page crashes when the logged in account have sightings
because the argument used for constructing the urls for all the bird links was
wrong. It was bird_id but should have been bird_binomial_name.

This has been corrected with this commit, and verified with a new integration
test.