Provide React with API URL through docker-compose
In order to be able to configure the dev environment and production environment
differently for the React frontend, we had to go through some loops. The
API_URL is now set in the docker-compose files (dev / prod) and upon starting
the containers, it is pulled from the environment variable, written to the file
env-config.js which is then put in the react public folder to be loaded in the
index.html head element.