Refactor and rework menu items
Menu items are now sent to navbar menu and sidebar menu as already constructed
components. That means that we should be able to customize better what will
happen when each item is clicked, like the logout link for example, which
should only trigger a logout, not navigate anywhere unless necessary.