Hook up React frontend email registration with API
Now, when the user provides their email address in the React frontend, to
initiate the registration proccess, the email is correctly sent to the backend
where the registration row is created in the database and an email is sent to
the provided email address. The frontend then displays a result alert message.