Implement password reset API POST resource
With this API resource, you can post a new password reset resource by providing
the email associated with the account you wish to reset the password to.

In order to make the links being sent out have the correct host (they are
different if the react frontend is being used or not) we had to create a flag
so that the FRONTEND_HOST is used instead of the Python Flask app
EXTERNAL_HOST.

Unit tests provided.