Update token on email registration conflict
In those instances when the user registers their email address again, we do not
care that there already was an unfinished registration in the database, but
instead just want to update the registration token and send out a new email to
the user. With this commit, we achieve just that.