Correctly stick footer at bottom of page
There was a problem when the content of the page was not enough to force the
footer down to its correct size at the bottom of the page, meaning that the
footer seemingly covered a vast amount of space on the page. With this commit,
this has been corrected, and the footer now has a fixed height and always appear
at the very bottom of the page.

References:
1. Fraser, D. "How to keep your footer where it belongs ?". freeCodeCamp.
   [Online] Available from:
   https://www.freecodecamp.org/news/how-to-keep-your-footer-where-it-belongs-59c6aa05c59c/
   [Accessed 28th of September 2019]