Store refresh token as JSON in local storage
It is extremely cumbersome and unnecessay to store three separate items in
local storage related to the refresh token. Instead, stringify the whole
refresh token into JSON and store and recover it as one single item.

Unfortunately, this means that users of the site prior to this commit will have
trash items left in the local storage forever unless we explicitly deletes
them. Therefore, as an attempt to clear some of this data, the app will delete
these items from local storage when it is starting up.

These three statements of clearing up items in local storage should be removed
after some time.