Implement versioning in frontend app
There is a problem in deploying new versions of the frontend app if previous
items in local storage have changed or become obsolete. This could cause
problems or the very least bloat the local storage. We need a general system of
telling a connecting user's browser, after such updates, to clear their local
storage and get a fresh start on the newest version of the app.

To do this, we now setup a hard coded app version that can be compared with
previously stored versions in local storage. On a discrepancy, the local
storage is wiped and the new version is stored instead.

Since the production and development environment differ, we will have to use
slightly different techniques to provide this version number to the docker
images/containers. In order to only have to change this version number in one
place in the future, we place it in the .env file that will get picked up by
the docker compose files and from there used differently.

- Production environment

For the production environment, the app version has to be provided during the
building of the docker image, because it has to be available in the step when
the production build of the React app is being put together. Thus, we provide
the Dockerfile the app version as an argument so that it can be converted into
an actual environment variable inside the image under construction. This
environment variable is then picked up by the env.sh script when the
env-config.js file is being put together and placed in the public folder of the
React app, before the app gets built.

- Development environment

In the development environment, the React app doesn't have to be built and
packaged into a production build, so it is enough to provide it as an
environment variable directly to the already finished built image, in the
docker compose, to be available when the image is finally being started (which
is the point when the env.sh script gets executed in development).