Rename AuthenticationContext to UserContext
This change comes for two reasons. First of all, the name is shorter. Secondly,
it is kind of more to the point. It is truly the context of the user of the
app. Is the user authenticated? What's their account? What's their access
token? It has nothing to do with the actual authentication process, but more of
the context once a user has finally become authenticated.