Show spinner while loading username list
It doesn't look nice when navigating to the home page while logged in, and you
get no indication that it is loading anything (the username list) which then
suddenly pops into existance out of nowhere. Now, we keep a loading state in
the home component and show the spinner while its true, and only set it to
false once the usernames have been fetched.