Correct layout width stretching issues
In some cases, the content that is being rendered on the page is too wide
which causes the whole page to stretch its width to accomodate for the content,
instead of pushing the content together. This causes the navbar to suddenly
stretch out of view and a horizontal scroll bar to appear, which looks very
bad.

To fix this isseu, we had to adjust some widths and overflow properties in the
CSS.