Modestly start using SQLAlchemy for Birder table
It is becoming quite hard to write the more and more complex SQL queries by
hand, and mixing query languages and Python makes the source code messy. I've
been introduced to the concept of ORM which will speed up development and the
code quality significantly. For this, I will be using SQLAlchemy.