Rid Flask app of Database database dependency
Instead, the database dependency, which is still needed for some tests to run,
is created in the AppTestCase set up. And furthermore, an IntegrationTestCase
class have now been added that will skip the included tests if DATABASE_HOST is
missing from the environment variable. This makes it possible to run just the
non-integration unit tests outside the docker containers.