Begin versioning the API, starting with birds
It is good practice to version your APIs, so I've decided to start versioning
the Aveslog API, even while in development. The version I'll start with will be
v0, which will only be used until I feel somewhat happy with the complete API
by which point I will mark it as v1 and be more careful about any changes I
make to it. By now, however, there are no guarantees at all, and things will be
very volatile.

Not all routes are versioned yet, since I wanted to start just with one, to
figure out the structure and how to do it. For simplicity's sake, you don't
have to prefix the routes with '/v0', but you may, if you want.