Rate limit and proxy fix remote address
We want to rate limit our API, which we achieve by using Flask-Limiter, but in
order to limit on the correct remote address, and not the Nginx proxy's remote
address, used in production, we have to try to make use of  the X-Forwarded-For
set by the proxy. Supposedly, it will just work magically if we use the
ProxyFix middleware by Werkzeug.

So, now, if BEHIND_PROXY is set and true in the environment variables for the
API application, it will use the proxy fix.

The response also includes the rate limit headers.