Add roles and permissions to REST API
Permissions are rules allowing certain HTTP methods to resources matching a
spceified regex. So for example, allowing modification of birds and all of
their subresources would be a permission with method POST and resource_regex
"^/birds/.*$"

Permissions are assigned to roles. And roles are assigned to accounts. So, for
example, we could have a "superadmin" role which contain the permission to use
POST on any resource matching the regex "^/.*$" (any resource). Dangerous?
Myes, but with this new role and permissions system it is atleast doable.