Adjust CommonNameItem to properly capitalize names
Separating the language code section from the name section by only using span
elements meant that the first word of the name didn't get capitalized, because
it was wrongly assumed to be just a continuation of the langauge code.

With this commit, the language code and the name section are now in two
different divs.