Stop using scale together with transform
The scale CSS property seem to only have an effect in Firefox 60+, so on
Firefox 60+, both the transform and the scale properties seems to be applied
to scale the bird thumbnail picture twice, which is obviously wrong. Since
scale is only a convenience property, this commit will stop using it in favor
of only using transform (with webkit-transform), which works as expected across
all browsers.