Correct brand name font for Android Firefox
For some reason, Firefox on Android doesn't want to correctly use the Courier
font family, or monospace like any other browser, and thus the brand name looks
like crap. For this reason, we now include a custom Courier font family and use
that one specifically, instead of letting the browser choose a font.