Use correct regex for username
There was a bug where the frontend allowed uppercase letters in the username,
but the backend did not. This caused errors when trying to register a new
account without any indication on why it didn't work. With this commit, the
frontend will now use the same regex as the backend.