Scroll restore page on every pathname change
There is an issue on the page when you move between pages that both have tall
contents causing the newly navigated-to page to be shown already scrolled down
as much as the previous page was. This is annoying and a better way is to
restore the scroll position when the router pathname has changed.