Make unit tests more performant
Of course it is nice if we recreate the app environment for each individual
test, but the tests are getting slow, and slow tests are tests that aren't run.
So to fix this, we now create one test_app and reuse that in all tests instead.
This cut the testing time in half. Greater care has to be taken making sure
that the tests doesn't affect each other, but it is worth it in my opinion.