Correct environment variable name for app version
I previously thought that the environment variables passed on to the React
application were stripped of their required REACT_APP_ prefix, meaning that if
I wanted to access APP_VERSION, the environment variable would have to be
REACT_APP_APP_VERSION, and then acessible just through APP_VERSION.

However, this is NOT the case. The prefixes are not removed, meanining that I
still need to refer to it as REACT_APP_APP_VERSION even from within the React
app.