Don't use old access token on retried request
When an access token expires, the axios response interceptor notices this and
requests a new access token. After this, it retries the initial request, that
first failed. However, the request interceptor were still putting the old
access token into the retried request. So with this commit, apart from doing
some refactoring, we also make sure to only put the access token into requests
through the request interceptor if they're not retry requests.